add login and logout button, show navbar in home page
