+++ src/app/[locale]/page.tsx
@@ -2,12 +2,12 @@ import Image from "next/image";
 import { getServerSession } from "next-auth/next";
 import Link from "~/components/link";
 import { useTranslations } from "next-intl";
-import LocaleSelect from "~/components/locale-select";
-import LoginButton from "~/components/login-button";
 import { type Session } from "next-auth";
+import { authOptions } from "~/server/auth";
+import Navbar from "~/components/navbar";
 
-export default async function Home() {
-  const session = await getServerSession();
+export default async function Page() {
+  const session = await getServerSession(authOptions);
 
   return <Content session={session} />;
 }
@@ -37,8 +37,7 @@ function Content({ session }: { session: Session | null }) {
         {t("page_home_quick_start_button")}
       </Link>
       <Link href="/setup">{t("page_home_custom_game_button")}</Link>
-      <LocaleSelect />
-      {!session && <LoginButton />}
+      <Navbar />
     </div>
   );
 }
